Move VersionedProperty to its own file. (#647)
